Initial 2.0 refactor
