+++ pytiled_parser/tiled_map.py
@@ -1,19 +1,12 @@
-# pylint: disable=too-few-public-methods
-
-import json
 from pathlib import Path
-from typing import Dict, List, Optional, Union
-from typing import cast as typing_cast
+from typing import Dict, List, Optional
 
 import attr
-from typing_extensions import TypedDict
 
-from . import layer, properties, tileset
-from .common_types import Color, Size
-from .layer import Layer, RawLayer
-from .properties import Properties, RawProperty
-from .tileset import RawTileSet, Tileset
-from .util import parse_color
+from pytiled_parser.common_types import Color, Size
+from pytiled_parser.layer import Layer
+from pytiled_parser.properties import Properties
+from pytiled_parser.tileset import Tileset
 
 TilesetDict = Dict[int, Tileset]
 
@@ -68,146 +61,3 @@ class TiledMap:
     hex_side_length: Optional[int] = None
     stagger_axis: Optional[str] = None
     stagger_index: Optional[str] = None
-
-
-class _RawTilesetMapping(TypedDict):
-    """ The way that tilesets are stored in the Tiled JSON formatted map."""
-
-    firstgid: int
-    source: str
-
-
-class _RawTiledMap(TypedDict):
-    """The keys and their types that appear in a Tiled JSON Map.
-
-    Keys:
-        compressionlevel: not documented - https://github.com/bjorn/tiled/issues/2815
-    """
-
-    backgroundcolor: str
-    compressionlevel: int
-    height: int
-    hexsidelength: int
-    infinite: bool
-    layers: List[RawLayer]
-    nextlayerid: int
-    nextobjectid: int
-    orientation: str
-    properties: List[RawProperty]
-    renderorder: str
-    staggeraxis: str
-    staggerindex: str
-    tiledversion: str
-    tileheight: int
-    tilesets: List[_RawTilesetMapping]
-    tilewidth: int
-    type: str
-    version: Union[str, float]
-    width: int
-
-
-def parse_map(file: Path) -> TiledMap:
-    """Parse the raw Tiled map into a pytiled_parser type
-
-    Args:
-        file: Path to the map's JSON file
-
-    Returns:
-        TileSet: a properly typed TileSet.
-    """
-
-    with open(file) as map_file:
-        raw_tiled_map = json.load(map_file)
-
-    parent_dir = file.parent
-
-    raw_tilesets: List[Union[RawTileSet, _RawTilesetMapping]] = raw_tiled_map[
-        "tilesets"
-    ]
-    tilesets: TilesetDict = {}
-
-    for raw_tileset in raw_tilesets:
-        if raw_tileset.get("source") is not None:
-            # Is an external Tileset
-            tileset_path = Path(parent_dir / raw_tileset["source"])
-            with open(tileset_path) as raw_tileset_file:
-                tilesets[raw_tileset["firstgid"]] = tileset.cast(
-                    json.load(raw_tileset_file),
-                    raw_tileset["firstgid"],
-                    external_path=tileset_path.parent,
-                )
-        else:
-            # Is an embedded Tileset
-            raw_tileset = typing_cast(RawTileSet, raw_tileset)
-            tilesets[raw_tileset["firstgid"]] = tileset.cast(
-                raw_tileset, raw_tileset["firstgid"]
-            )
-
-    if isinstance(raw_tiled_map["version"], float):
-        version = str(raw_tiled_map["version"])
-    else:
-        version = raw_tiled_map["version"]
-
-    # `map` is a built-in function
-    map_ = TiledMap(
-        map_file=file,
-        infinite=raw_tiled_map["infinite"],
-        layers=[layer.cast(layer_, parent_dir) for layer_ in raw_tiled_map["layers"]],
-        map_size=Size(raw_tiled_map["width"], raw_tiled_map["height"]),
-        next_layer_id=raw_tiled_map["nextlayerid"],
-        next_object_id=raw_tiled_map["nextobjectid"],
-        orientation=raw_tiled_map["orientation"],
-        render_order=raw_tiled_map["renderorder"],
-        tiled_version=raw_tiled_map["tiledversion"],
-        tile_size=Size(raw_tiled_map["tilewidth"], raw_tiled_map["tileheight"]),
-        tilesets=tilesets,
-        version=version,
-    )
-
-    layers = [layer for layer in map_.layers if hasattr(layer, "tiled_objects")]
-
-    for my_layer in layers:
-        for tiled_object in my_layer.tiled_objects:  # type: ignore
-            if hasattr(tiled_object, "new_tileset"):
-                if tiled_object.new_tileset:
-                    already_loaded = None
-                    for val in map_.tilesets.values():
-                        if val.name == tiled_object.new_tileset["name"]:
-                            already_loaded = val
-                            break
-
-                    if not already_loaded:
-                        highest_firstgid = max(map_.tilesets.keys())
-                        last_tileset_count = map_.tilesets[highest_firstgid].tile_count
-                        new_firstgid = highest_firstgid + last_tileset_count
-                        map_.tilesets[new_firstgid] = tileset.cast(
-                            tiled_object.new_tileset,
-                            new_firstgid,
-                            tiled_object.new_tileset_path,
-                        )
-                        tiled_object.gid = tiled_object.gid + (new_firstgid - 1)
-
-                    else:
-                        tiled_object.gid = tiled_object.gid + (
-                            already_loaded.firstgid - 1
-                        )
-
-                    tiled_object.new_tileset = None
-                    tiled_object.new_tileset_path = None
-
-    if raw_tiled_map.get("backgroundcolor") is not None:
-        map_.background_color = parse_color(raw_tiled_map["backgroundcolor"])
-
-    if raw_tiled_map.get("hexsidelength") is not None:
-        map_.hex_side_length = raw_tiled_map["hexsidelength"]
-
-    if raw_tiled_map.get("properties") is not None:
-        map_.properties = properties.cast(raw_tiled_map["properties"])
-
-    if raw_tiled_map.get("staggeraxis") is not None:
-        map_.stagger_axis = raw_tiled_map["staggeraxis"]
-
-    if raw_tiled_map.get("staggerindex") is not None:
-        map_.stagger_index = raw_tiled_map["staggerindex"]
-
-    return map_
