+++ pytiled_parser/version.py
@@ -1,3 +1,3 @@
 """pytiled_parser version"""
 
-__version__ = "1.5.4"
+__version__ = "2.0.0-beta"
