+++ setup.cfg
@@ -44,7 +44,7 @@ tests =
   pytest
   pytest-cov
   black
-  pylint
+  flake8
   mypy
   isort<5,>=4.2.5
 
@@ -104,3 +104,7 @@ strict_optional = True
 
 [mypy-tests.*]
 ignore_errors = True
+
+[flake8]
+max-line-length = 88
+exclude = .git,__pycache__,__init__.py,.mypy_cache,.pytest_cache
