+++ tests/test_layer.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 import pytest
 
-from pytiled_parser import layer
+from pytiled_parser.parsers.json.layer import parse
 
 TESTS_DIR = Path(os.path.dirname(os.path.abspath(__file__)))
 TEST_DATA = TESTS_DIR / "test_data"
@@ -39,6 +39,6 @@ def test_layer_integration(layer_test):
 
     with open(raw_layers_path) as raw_layers_file:
         raw_layers = json.load(raw_layers_file)["layers"]
-        layers = [layer.cast(raw_layer) for raw_layer in raw_layers]
+        layers = [parse(raw_layer) for raw_layer in raw_layers]
 
     assert layers == expected.EXPECTED
